+++ src/main/java/com/sweetmimike/perfectmobfarm/block/entity/GoldMobFarmEntity.java
@@ -1,6 +1,6 @@
 package com.sweetmimike.perfectmobfarm.block.entity;
 
-import com.sweetmimike.perfectmobfarm.config.CommonConfigs;
+import com.sweetmimike.perfectmobfarm.config.ServerConfigs;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
@@ -12,7 +12,7 @@ import net.minecraft.world.level.block.state.BlockState;
 public class GoldMobFarmEntity extends IronMobFarmEntity {
 
     public GoldMobFarmEntity(BlockPos pWorldPosition, BlockState pBlockState) {
-        super(BlockEntityManager.GOLD_MOB_FARM_ENTITY.get(), pWorldPosition, pBlockState, CommonConfigs.GOLD_MOB_FARM_COOLDOWN.get());
+        super(BlockEntityManager.GOLD_MOB_FARM_ENTITY.get(), pWorldPosition, pBlockState, ServerConfigs.GOLD_MOB_FARM_COOLDOWN.get());
     }
 
     @Override